Work Completed:
Implemented AsyncTaskLoader.
Continued development of NetworkUtils.
Added Permissions to the AndroidManifest.

<uses-permission android:name="android.permission.INTERNET" />
Tested with my API Key.

"page":1,"total_results":19826,"total_pages":992,"results":[{ ... }]
It would appear I was able to poll for data, and now I just need to work on parsing my results into a more proper structure.
Issues:

Network Calls on the main thread.

This is so discouraged it creates an exception. The delays of invoking a network call could impact the user experience, so the API outright prevents it.
Permissions are important! Don't forget to include the necessary permissions in the AndroidManifest.xml file.
<uses-permission android:name="android.permission.INTERNET" />
Possible Issue: resolving the suspected memory leak caused by dynamically creating an AsyncTaskLoader within the onCreateLoader call. (IDE identified this as a possible issue).