Popular Movies, Stage 2 - Resubmitted for Review on Day 76 (Today):
Feedback Received: 5 SPECIFICATIONS REQUIRE CHANGES
#1: Your app fails on this point “Maintains list items positions on device rotation.”.
Result: Numerous links were provided as reference... However, one link didn't work, while the other two related to implementation types that weren't compatible with my implementation. That said, despite the confusion I was able to troubleshoot this and find a workable solution.

#2: When a user changes the sort criteria (most popular, highest rated, and favorites) the main view gets updated correctly.
Result: the reviewer identified a bug, but didn't communicate how it occurred or other steps to reproduce. It turns out, it was an issue with the back buttons behaving differently from each other (action bar vs OS back button). Further, the AsyncLoader was loading duplicate items, so even when the rest of the data was up to date, the loader was repopulating the UI with old data.

#3: After I unfavorite the movie and go back to the list, it still shows that movie listed as the favorite. The view should be updated.
Result: This one tied (sort of...) to the above issue; different back buttons yielded different outcomes.

#4: If Room is used, database is not re-queried unnecessarily. LiveData is used to observe changes in the database and update the UI accordingly.
#5: If Room is used, database is not re-queried unnecessarily after rotation. Cached LiveData from ViewModel is used instead.
Result: This was completely my bad... I neglected to utilize LiveData, and hand-implemented my own solution initially. This has been corrected in the current version. Tht said, ViewModel is implied in AppCompatActivity, and apparently is not a hard-requirement of LiveData... I suspect my project may be rejected for this as it is a 'literal' rubric requirement, despite matching all functional criteria.

And now... Onto the code!

FavoriteUtils.java
*Stored a reference to the MainActivity context.
*Converted favoriteMovies into MutableLiveData.
*Added Observer class for monitoring LiveData.
*Added movieAdapter, to have the Observer update the adapter directly.
*Inadvertently re-added void to the constructor.
*Updated add & remove functions to account for MutableLiveData functionality.
*Added helper methods for Observer and adapters.

MainActivity.java
*Abstracted out the movieAdapter and gridView, from function-local scope to class scope, to allow them to be more accessible between database movies and network movies.
*Added KeyScrollPosition string to identify the new integer being stored that retains scroll position on rotation.
*Adjusted the loadFromDatabaseOrNetwork initial call, to load based on the stored preference selection.
*Added onPause override to retrieve the current scroll position prior to destruction.
*Overhauled onSharedPreferenceChanged during review/debugging, included check to ensure we weren't changing UI elements when the UI scroll position was saved.
*Destroys loader when loading from database, to prevent it from clobbering local data with previous calls.
*Updating scroll position after loading data.
*Created helper function to setup the adapter after loading data over the network.
*Added more logging throughout.

SettingsActivity.java
*Return true when the item is found.

SettingsFragment.java
*Added additional logging.

AndroidManifest.xml
*Accidental whitespace, I'll keep it for the luls.

build.gradle
*Added Necessary dependencies for ViewModel, LiveData, and LifeCycle.